# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Librarie API",
        "version": "0.1"
    },
    "paths": {
...
        "/carti": {
            "get": {
                "tags": [
                    "Carti"
                ],
                "summary": "Returneaza toate cartile",
                "description": "Returneaza toate cartile",
                "operationId": "4e78798be6b1bac63221ec7cc4cdaefb",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Carti"
                ],
                "summary": "Update cartea cu id-ul dat",
                "description": "Update cartea cu id-ul dat",
                "operationId": "f0b1a389e2ae435bc9d566a2aec3abba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "titlu": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "autor": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "imagine": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "isbn": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "editura": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "pret": {
                                        "description": "",
                                        "type": "float",
                                        "example": 1
                                    },
                                    "an_aparitie": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nr_pg": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tip_coperta": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "limba": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "dimensiune": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "cantitate": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Carti"
                ],
                "summary": "Adauga cartea",
                "description": "Adauga cartea",
                "operationId": "71a08f3c6c1b5ac8d7b87dc5fa0fdadd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titlu": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "autor": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "imagine": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "isbn": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "editura": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "pret": {
                                        "description": "",
                                        "type": "float",
                                        "example": 1
                                    },
                                    "an_aparitie": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nr_pg": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tip_coperta": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "limba": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "dimensiune": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "cantitate": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/carti/{id}": {
            "get": {
                "tags": [
                    "Carti"
                ],
                "summary": "Returneaza cartea cu id-ul dat",
                "description": "Returneaza cartea cu id-ul dat",
                "operationId": "b1cfd8a9a1405163e190d1798baa0ae5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul cartii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carti"
                ],
                "summary": "Sterge cartea cu id-ul dat",
                "description": "Sterge cartea cu id-ul dat",
                "operationId": "0eb0381ef598023ed1f8d4c65b33b511",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul cartii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/carti/categorii/{id}": {
            "get": {
                "tags": [
                    "Carti"
                ],
                "summary": "Returneaza categoriile cartii cu id-ul dat",
                "description": "Returneaza categoriile cartii cu id-ul dat",
                "operationId": "d9b64a7c0baf975f7a9707dabef0cca8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul cartii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/carti/reviewuri/{id}": {
            "get": {
                "tags": [
                    "Carti"
                ],
                "summary": "Returneaza reviewurile cartii cu id-ul dat",
                "description": "Returneaza reviewurile cartii cu id-ul dat",
                "operationId": "0d3b78ca8c122436a2addf13cba43b8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul cartii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}